Unify GUI features and simplify layout
