+++ pen_plotter/__init__.py
@@ -3,7 +3,7 @@
 from .font_paths import FontLoader, LayoutSettings, layout_text
 from .gcode import PlotterSettings, paths_to_gcode
 from .geometry import measure_paths, translate_paths
-from .gui import PlotterGUI
+from .gui import PenPlotterStudio, PlotterGUI, main
 
 __all__ = [
     "FontLoader",
@@ -13,5 +13,7 @@ __all__ = [
     "paths_to_gcode",
     "measure_paths",
     "translate_paths",
+    "PenPlotterStudio",
     "PlotterGUI",
+    "main",
 ]
